Adds MVC middleware for Prometheus Exporter.
Refactors the writing of metrics to a common utility class in
Implementation.

Removes the `assembly: System.CLSCompliant(true)]` directive from
OpenCensus.Exporter.Prometheus to allow support for non-CLS  code (which
are the MVC types).

Adds a test/example Application TestPrometheusMiddleware.AspNetCore.2.0
which shows the middleware working in an example MVC application.